pyproject.toml
[project]
name = "pyscript-bad-apple"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.1",
    "pillow>=11.1.0",
    "pyscript>=0.3.3",
]
